Refactor onboarding API and photo upload for pixel art avatars
- Replace previous onboarding route with new implementation supporting
roles and social links - Add badge generation including optimized text
rendering and QR codes - Refactor photo upload to generate 8-bit pixel
art avatars with background removal - Remove combined AI avatar and
badge generation; separate pixel art generation pipeline - Improve
database schema formatting and consistency
